Truthsocialapi/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.1
click>=8.1.0
loguru>=0.7.0
python-dateutil>=2.9.0
curl_cffi>=0.13.0
